requirements.txt
beautifulsoup4==4.12.3
botocore==1.34.55
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
starkbank-ecdsa==2.2.0
uritemplate==4.1.1
urllib3==1.26.18
